Add files via upload
This file presents the aggregated survey results from 39 respondents. There are 4 separate sheets in this file:

Demographics: This presents the country, language, and experience breakdown of the participants.

Collaboration: This quantifies the level of national and international collaboration, as well as awareness of various sources of MSS research.

Keywords: This measures the use of keywords in identifying research, and the preferences of MSS researchers to two types of keywords: 'Support' Keywords and Additional Keywords.

Terminology: This describes participants' attitudes towards the level of standardisation in MSS terminology, and their use or awareness of various terms under the themes identifed in the Thematic Analysis in Section 4.2: Physical Centres, Personnel, Methods, Research Topics, Tools/Techniques.